+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Feature: Create `ButtonGroup` component.
 - Feature: `ScrollableContainer` new props `shadowColor`, `backgroundColor`
+- Fix: `StickyContainer` fix for `isSticky='always'`
 
 ## [2.3.0] - 2019-04-19